Move user workflow code to TaskService
This was originally in `UserService`, but is specific to Tasks.
